SITES-21278: Code review Changes
